Initial Files
When I created the repository, it seems like that I failed. Today when I try to find them in my repository. They have gone! Fortunately, I find some files in my PC. But I can not be sure that it's the latest and I don't have vb6.0 in my PC. I will check them and finish it in summer vocation at the end of July at home.